Reduced css lines & username button fix. Fixes a lot of front end, seperating a lot of logic and removing them in respective file while importing shared and common element styling to "style.css".
Changes:
- Style.css: handles variables alongside common styling approaches, thus reducing a lot of unnecessary lines in other files.
- *.html: for some removed unnecessary classes and improved semantics, in addition for some added "%%usernameButton%%" template indicator.
New file:
- username-button.css for styling specific for very reoccuring html element in a few files (home, post-view).
Removed:
- login.css: as I renamed "signup.css" to "signup-login.css" as I noticed it was just the same (more or less) file.

Backend:
- utilMethods.js: Moved function replaceDangerousChars here and new function for helping templating for login or logged in user view.
- userhandler.js, homeHandler.js, postHandler.js: updated accordingly for above changes.
