+++ private/templates/home.html
@@ -1,48 +1,46 @@
 <!DOCTYPE html>
 <html lang="en">
+
 <head>
     <link rel="preconnect" href="https://fonts.googleapis.com"> <!--Font From Google Fonts-->
     <link rel="preconnect" href="https://fonts.gstatic.com" crossorigin>
     <link href="https://fonts.googleapis.com/css2?family=Tektur:wght@400..900&display=swap" rel="stylesheet">
     <link rel="preconnect" href="https://fonts.googleapis.com">
     <link rel="preconnect" href="https://fonts.gstatic.com" crossorigin>
-    <link href="https://fonts.googleapis.com/css2?family=Josefin+Sans:ital,wght@0,100..700;1,100..700&display=swap" rel="stylesheet">
+    <link href="https://fonts.googleapis.com/css2?family=Josefin+Sans:ital,wght@0,100..700;1,100..700&display=swap"
+        rel="stylesheet">
     <meta charset="UTF-8">
     <meta name="viewport" content="width=device-width, initial-scale=1.0">
+    <link rel="stylesheet" href="/public/css/style.css">
     <link rel="stylesheet" href="/public/css/home.css">
+    <link rel="stylesheet" href="/public/css/username-button.css">
     <script defer src="/public/js/like.js"></script>
+    <script defer type="module" src="/public/js/detectSwipe.js"></script>
     <title>Document</title>
 </head>
-<body>
-  <header class="topRow">
-    <form action="/search" method="get">
-    <input class="search" name= "q" type="text" spellcheck="true" maxlength="100"
-    placeholder="Search">
-    <form action="/user/username">
-      <button id="standardButton" onclick=""> 
-        %%Username%%
-      </button>
-    </form>
-    </form>
 
-    <a href="/user/username"> 
-      <div class="profile"></div>
-    </a>
-  </header>
-  <main>
-    <section class="post-board">
-      %%posts%%
-    </section>
-  </main>
-  <div id="profilePanel" class = "hidden"></div>
-  <script type ="module" src="/public/js/detectSwipe.js"></script>
-  <footer class="footer">
-    <h2 class="title">
-      Contact
-    </h2>
-    <a class="github" href="https://github.com/SonnyVincenzo/SoftwareDevelopment_Project_NodeJs_SocialMedia">
-      Github
-    </a>
-  </footer>
+<body>
+    <header>
+        <form action="/search" method="get">
+            <input class="search" name="q" type="text" spellcheck="true" maxlength="100" placeholder="Search">
+        </form>
+
+        %%usernameButton%%
+    </header>
+    <main>
+        <section class="post-board">
+            %%posts%%
+        </section>
+    </main>
+    <div id="profilePanel" class="hidden"></div>
+    <footer>
+        <h2 class="title">
+            Contact
+        </h2>
+        <a class="github" href="https://github.com/SonnyVincenzo/SoftwareDevelopment_Project_NodeJs_SocialMedia">
+            Github
+        </a>
+    </footer>
 </body>
+
 </html>
