+++ private/templates/index.html
@@ -2,58 +2,60 @@
 <html lang="en">
 
 <head>
-  <link rel="preconnect" href="https://fonts.googleapis.com"> <!-- Taken from Google Fonts -->
-  <link rel="preconnect" href="https://fonts.gstatic.com" crossorigin>
-  <link
-    href="https://fonts.googleapis.com/css2?family=Josefin+Sans:ital,wght@0,100..700;1,100..700&family=Tektur:wght@400..900&display=swap"
-    rel="stylesheet">
-  <meta charset="UTF-8">
-  <meta name="viewport" content="width=device-width, initial-scale=1.0">
-  <link rel="stylesheet" href="/public/css/index.css">
-  <script defer src="/public/js/indexSearchBar.js"></script>
-  <title>Welcome!</title>
+    <link rel="preconnect" href="https://fonts.googleapis.com"> <!-- Taken from Google Fonts -->
+    <link rel="preconnect" href="https://fonts.gstatic.com" crossorigin>
+    <link
+        href="https://fonts.googleapis.com/css2?family=Josefin+Sans:ital,wght@0,100..700;1,100..700&family=Tektur:wght@400..900&display=swap"
+        rel="stylesheet">
+    <meta charset="UTF-8">
+    <meta name="viewport" content="width=device-width, initial-scale=1.0">
+    <link rel="stylesheet" href="/public/css/style.css">
+    <link rel="stylesheet" href="/public/css/index.css">
+    <script defer src="/public/js/indexSearchBar.js"></script>
+    <title>Welcome!</title>
 
 </head>
 
 <body>
-  <header class="header">
-    <h1 class="title"> The Name of This Thing </h1>
-    <a class="homeButton" href="/home">Home</a>
-  </header>
+    <header>
+        <h1 class="title"> The Name of This Thing </h1>
+        <a class="homeButton" href="/home">Home</a>
+    </header>
 
-  <main>
-    <section class="searchBar">
-      <p class="searchText">Search |</p>
-      <form method="GET" action="/search">
-      <input id="searchInput" name ="q" type="text" spellcheck="true" maxlength="80" placeholder="Post's, Username">
-      <span class="clearButton" onclick="clearInput()"></span>
-      </form>
-    </section>
-    <div class="standardButton">
-      <a class="text" href="/auth/login">Login</a>
-      <span>
-        <p class="text"> | </p>
-      </span>
-      <a class="text" href="/auth/signup">Create Account</a>
-    </div>
+    <main>
+        <section class="searchBar">
+            <p class="searchText">Search |</p>
+            <form method="GET" action="/search">
+                <input id="searchInput" name="q" type="text" spellcheck="true" maxlength="80"
+                    placeholder="Post's, Username">
+                <span class="clearButton" onclick="clearInput()"></span>
+            </form>
+        </section>
+        <div id="user-options">
+            <a class="text" href="/auth/login">Login</a>
+            <span>
+                <p class="text"> | </p>
+            </span>
+            <a class="text" href="/auth/signup">Create Account</a>
+        </div>
 
-  </main>
+    </main>
 
-  <footer class="footer">
-    <div class="footer-container">
-      <h2 class="footerTitle">Contact</h2>
-      <a class="github" href="https://github.com/SonnyVincenzo/SoftwareDevelopment_Project_NodeJs_SocialMedia">
-        Github
-      </a>
-    </div>
-    <div class="footer-container">
-      <h2 class="footerTitle">Easter Egg</h2>
-      <a class="github" href="/post-feed">
-        post-feed
-      </a>
-    </div>
-  </footer>
-  
+    <footer>
+        <div class="footer-container">
+            <h2 class="footerTitle">Contact</h2>
+            <a class="github" href="https://github.com/SonnyVincenzo/SoftwareDevelopment_Project_NodeJs_SocialMedia">
+                Github
+            </a>
+        </div>
+        <div class="footer-container">
+            <h2 class="footerTitle">Easter Egg</h2>
+            <a class="github" href="/post-feed">
+                post-feed
+            </a>
+        </div>
+    </footer>
+
 </body>
 
 </html>
