+++ private/templates/user.html
@@ -7,18 +7,18 @@
     <link href="https://fonts.googleapis.com/css2?family=Josefin+Sans:ital,wght@0,100..700;1,100..700&display=swap" rel="stylesheet">
     <meta charset="UTF-8">
     <meta name="viewport" content="width=device-width, initial-scale=1.0">
+    <link rel="stylesheet" href="/public/css/style.css">
     <link rel="stylesheet" href="/public/css/user.css">
     <script defer src="/public/js/like.js"></script>
     <title>Profile Page</title>
 </head>
 
 <body>
-    <div class="header">
+    <header>
         <a class="aButton" href="/home">Home</a>
-    </div>
+    </header>
 
-    <div class="mainContainer">
-        
+    <main>
         <div class="profile-column">
             <div class="profile"></div> <div class="profileName">
                 <h1>%%fullName%%</h1>
@@ -46,6 +46,6 @@
                 %%posts%%
             </article>
         </div>
-    </div> 
+    </main> 
 </body>
 </html>
