--- public/css/login.css
@@ -1,112 +1,0 @@
-:root {
-    --primary-color: rgb(43, 42, 43);
-    --secondary-color: rgb(138, 43, 226);
-    --accent-color: rgb(1, 200, 250); 
-}
-
-body {
-    background-image: url(/public/assets/img/asfalt-light.png); /*Background Texture found online*/
-    background-color: var(--primary-color);
-    margin: 0;
-    font-family: 'Tektur', sans-serif;
-    display: flex;
-    flex-direction: column;
-    min-height: 100vh;
-
-}
-
-nav {
-    text-align: center;
-    padding: 15px;
-    background-color: var(--primary-color);
-    border-bottom: 1px solid var(--secondary-color);
-}
-
-.signup-container {
-    flex: 1;
-    display: flex;
-    flex-direction: column; 
-    align-items: center;    
-    justify-content: flex-start; 
-    padding-top: 100px;
-    width: 100%;
-}
-form {
-    background-color: var(--primary-color);
-    border: 2px solid var(--accent-color);
-    padding: 40px;
-    border-radius: 12px;
-    width: 100%;
-    max-width: 400px;
-    display: flex;
-    flex-direction: column;
-    gap: 20px; 
-}
-
-.username, .password, .dob {
-    margin-bottom: 25px;
-    display: flex;
-    flex-direction: column;
-}
-
-input {
-    padding: 12px;
-    border: 1px solid #555;
-    border-radius: 5px;
-    color: black;
-}
-
-label {
-    color: white;
-}
-
-
-.standardButton {
-    background-color: var(--primary-color);
-    border: 2px solid violet;
-    border-radius: 5px;
-    color: white;
-    cursor: pointer;
-    font-family: 'Tektur', sans-serif;
-    font-size: 18px;
-    padding: 10px 20px;
-    font-weight: bold;
-    margin-top: 10px;
-    transition: 0.3s;
-}
-
-.standardButton:hover {
-    background: linear-gradient(to right, var(--secondary-color), rgba(1, 199, 249, 0.8));
-    color: black;
-    border-color: white;
-}
-
-.text {
-    color: white;
-    font-family: 'Josefin Sans', sans-serif;
-    font-size: 14px;
-    margin: 5px;
-    padding-top: 5px;
-}
-
-.popup-overlay {
-    position: fixed;
-    top: 0;
-    left: 0;
-    width: 100%;
-    height: 100%;
-    background: rgba(0, 0, 0, 0.5);
-    display: flex;
-    justify-content: center;
-    align-items: center;
-    z-index: 999;
-}
-
-.popup-content {
-    background: #fff;
-    padding: 24px;
-    border-radius: 8px;
-    text-align: center;
-    box-shadow: 0 4px 12px rgba(0, 0, 0, 0.15);
-    max-width: 350px;
-}
