+++ public/css/post-view.css
@@ -3,15 +3,7 @@ html, body {
   margin: 0px
 }
 
-html{
-    background-color: rgb(43, 42, 43);
-    background-image: url(/public/assets/img/asfalt-light.png); /*Background Texture found online*/
-}
 
-body {
-  display: flex;
-  flex-direction: column;
-}
 main{
   flex: 1;
 }
@@ -39,7 +31,7 @@ main{
   margin: 0px;
   background-color: rgba(53, 52, 53, 0.8);
   height: 90px;
-  border-bottom: 1px rgba(125, 124, 124, 0.933) solid;
+  border-bottom: var(--header-footer-border-divider);
   justify-content: right;
 
   gap: 20px;
@@ -61,9 +53,9 @@ main{
   outline:none;
 }
 
-#standardButton{
+.standardButton{
   display: inline-block;
-  background-color: rgb(43, 42, 43);
+  background-color: var(--primary-bg-color);
   border-width: 3px;
   border-color: blueviolet;
   border-radius: 5px;
@@ -121,7 +113,7 @@ main{
   border-width: 2px;
   border-color: rgba(1, 199, 249, 1);
   border-radius: 10px;
-  background-color:rgb(43, 42, 43);
+  background-color:var(--primary-bg-color);
   width: 100%;
 }
 
@@ -142,8 +134,8 @@ main{
 .dislike,
 .like{
   display: flex ;
-  color: rgb(43, 42, 43);
-  background-color:rgb(43, 42, 43);
+  color: var(--primary-bg-color);
+  background-color:var(--primary-bg-color);
   font-size: 28px;
   border: none;
   -webkit-text-stroke: 2px blueviolet;
@@ -183,7 +175,7 @@ main{
 
 .aButton {
     display: inline-block;
-    background-color: rgb(43, 42, 43);
+    background-color: var(--primary-bg-color);
     justify-content: center;
     border: 2px solid violet;
     border-radius: 5px;
@@ -247,7 +239,7 @@ main{
   margin: 0px;
   background-color: rgba(53, 52, 53, 0.8);
   height: 90px;
-  border-top: 1px rgba(125, 124, 124, 0.933) solid;
+  border-top: var(--header-footer-border-divider);
 }
 
 .github{
