+++ public/css/signup-login.css
@@ -1,28 +1,4 @@
-:root {
-    --primary-color: rgb(43, 42, 43);
-    --secondary-color: rgb(138, 43, 226);
-    --accent-color: rgb(1, 200, 250); 
-}
-
-body {
-    background-image: url(/public/assets/img/asfalt-light.png); /*Background Texture found online*/
-    background-color: var(--primary-color);
-    margin: 0;
-    font-family: 'Tektur', sans-serif;
-    display: flex;
-    flex-direction: column;
-    min-height: 100vh;
-
-}
-
-nav {
-    text-align: center;
-    padding: 15px;
-    background-color: var(--primary-color);
-    border-bottom: 1px solid var(--secondary-color);
-}
-
-.signup-container {
+main {
     flex: 1;
     display: flex;
     flex-direction: column; 
@@ -31,8 +7,8 @@ nav {
     padding-top: 100px;
     width: 100%;
 }
-form {
-    background-color: var(--primary-color);
+main form {
+    background-color: var(--primary-bg-color);
     border: 2px solid var(--accent-color);
     padding: 40px;
     border-radius: 12px;
@@ -42,8 +18,16 @@ form {
     flex-direction: column;
     gap: 20px; 
 }
+form > a {
+    text-align: center;
+    color: white;
+    transition: ease-in-out 0.2s;
+}
+form > a:hover {
+    color: var(--accent-color);
+}
 
-.username, .password, .dob {
+.user-input {
     margin-bottom: 25px;
     display: flex;
     flex-direction: column;
@@ -60,9 +44,13 @@ label {
     color: white;
 }
 
+#tos + label:hover {
+    cursor: pointer;
+}
+
 
 .standardButton {
-    background-color: var(--primary-color);
+    background-color: var(--primary-bg-color);
     border: 2px solid violet;
     border-radius: 5px;
     color: white;
