+++ public/css/user.css
@@ -1,25 +1,6 @@
-:root {
-    --primary-color: rgb(43, 42, 43);
-    --secondary-color: rgb(138, 43, 226);
-    --accent-color: rgb(1, 200, 250); 
-}
-
-body {
-    background-color: var(--primary-color);
-    color: white;
-    margin: 0;
-    font-family: Tektur, sans-serif;
-}
-
-nav {
-    text-align: center;
-    padding: 15px;
-    background-color: var(--primary-color);
-    border-bottom: 1px solid var(--secondary-color);
-}
-
-.mainContainer {
+main {
     display: flex;
+    color: white;
     flex-direction: row;
     width: 100%;
     min-height: 100vh;
@@ -28,7 +9,7 @@ nav {
 .profile-column {
     flex: 1;
     padding: 40px 20px;
-    background-color: var(--primary-color);
+    background-color: var(--primary-bg-color);
     border-right: 1px solid var(--secondary-color);
     display: flex;
     flex-direction: column;
@@ -37,10 +18,6 @@ nav {
 }
 
 .profile {
-    display: inline-block;
-    background: radial-gradient(circle, rgba(1, 200, 250, 0.8), rgba(138, 43, 226, 0.8));
-    background-color: white;
-    border-radius: 25px;
     width: 120px;
     height: 120px;
     margin-bottom: 10px;
@@ -54,7 +31,6 @@ nav {
 .posts-column {
     flex: 3;
     padding: 20px;
-    background-image: url(/public/assets/img/asfalt-light.png); /*Background Texture found online*/
     overflow-y: auto;
 }
